+++ src/plugin/mod.rs
@@ -1,6 +1,15 @@
 pub use self::configuration::{RapierConfiguration, SimulationToRenderTime, TimestepMode};
-pub use self::context::RapierContext;
-pub use self::plugin::{NoUserData, PhysicsSet, RapierPhysicsPlugin, RapierTransformPropagateSet};
+pub use self::context::{
+    systemparams::{
+        DefaultRapierContextAccess, DefaultRapierContextAccessMut, RapierContextAccess,
+        RapierContextAccessMut,
+    },
+    DefaultRapierContext, RapierContext, RapierContextEntityLink,
+};
+pub use self::plugin::{
+    NoUserData, PhysicsSet, RapierContextInitialization, RapierPhysicsPlugin,
+    RapierTransformPropagateSet,
+};
 
 #[allow(clippy::type_complexity)]
 #[allow(clippy::too_many_arguments)]